remote_storage: move errors to error.rs
